Fix core1 stack pointer setup
Take the address of the _core1_estack symbol to get the correct
address of the end of the stack area where the stack pointer should
begin.

Change the parameters to core1_launch so that the values are simply
usize types since that is how they are most closely used. This
means less casting and cleaner calls.