Use static_init! to create HardwareSyncBlock
Hide the initialization details of HardwareSyncBlock behind the
initialization of HardwareSyncBlockAccess. Use a global instance
reference for SIOSpinlock to mark itself released after it is
dropped.

Remove the bare initialization and access functions to keep access
behind the access type.

This way, we can use static_init! for just one resource in the
board's main function and get both when code calls it for the first
time.

This makes HardwareSyncBlockAccess practically fieldless and cheap
to create.