Make using debug! multicore-safe-ish
Allow Tock to print debug messages from multiple cores without
wrecking state. However, due to the granularity being limited to
within the debug writer, separate calls can result in broken output
which is exactly what happens in debug! with one write for the
passed-in string and one for the new line.