Phase two of interrupt handling
Provide an interrupt handler to ignore and disable an interrupt (in
the case of the UART0 interrupt which I want core0 to handle). This
is the alternative to modifying the UART code to not enable the
interrupt if the executing core is core1 (doesn't feel like the
right solution).

Copy the IPSR into r0 whenever entering the interrupt handlers so
it is easier to see what interrupt got control flow here.