Get DMA reading from an active ADC
Write a hack into the ADC since there is no sophisticated HIL
support for ADC -> DMA transfers. This gets us a single successful
transmission and locked into the ignored_interrupt handler simply
looping after the DMA channel finishes the transfer.

Rework the buffer management to have sample collection handled in
the DSP loop. Since a DMA client must have a static lifetime, it
is not possible to not create a static client, which I want to
avoid at this point. Simply point the DMA channel over to the
engine itself and keep the buffers within the engine.

Define the basics of the effects chain and add a dummy no-op link
to use for testing for now.

Start work on using the UserspaceKernelBoundary to create a
context, which is a simpler, more general concept from creating a
process. Allocate some space in memory (provided by the linker
script) for the context's stack.

Fix an issue with the FIFO having an interrupt pending before Tock
even runs. This happens because of a read of the FIFO that core1
performs in the bootrom code.