Cycle through buffersr to collect samples
Use the interrupt handler to call to the DMA object to handle the
DMA interrupt. Swap in the next free buffer in the sequence during
the callback that DSPEngine gets from the DMA channel.

If we completely fill all buffers, perform a panic.

Rework the type definitions of the signal chain.