Use two buffers for processing
Provide signal processors two buffers to do the processing: one
for the provided samples and one for the new samples. This will
make processing easier instead of having to do in-place processing
which is more complex. This comes at the cost of possibly longer
processing times since doing this precludes any clever processing
on a single buffer.