Fix GPIO, PIO for basic example
Make some fixes to enable the PIO state machine the the caller
provides a configuration for. Also fix the configure() function on
the PIO struct to configure a PIO block once.

GPIO was previously working through the single-cycle I/O facilities
on the RP2040. Not entirely sure yet, but this may be inhibiting
getting the PIO working correctly.