Fix buffer allocation from aliasing
There is some strangeness with how static_buf behaves where it will
return the same memory address for subsequent calls to the same
function. Move the buffer allocation to the engine and pass the
allocated buffers to the container.

Now, there appears to be some instances where the difference in
timing between the source and sink causes one to catch up to the
other and cause a failure. We may be able to solve this by
dynamically adjusting the source/sink rates such that the source
never runs up to the sink and the sink never falls behind too
much.