Add implementations for DSP control
Modify the FIFO code to hold two clients, one for each core. The
FIFO selects the right client depending on the core the code is
executing on.

Add two new types implementing the DSP controller and command
receiver. Give the controller to the DSPControl capsule and give
the receiver to the DSP engine.