Move signal preprocessing to ProcessControl
Let each ProcessControl implementation handle any necessary
preprocessing on the incoming samples.

Use this to add two types of preprocessing to the ADC-to-I2S
implementation. One will use the 12-bit samples as is with just a
shift toward a baseline. The other will scale them up to 16 bits
and also shift them.

With the UDA1334, it seems that using 12-bit samples only results
in a smaller, corresponding waveform, yielding a noisy signal. The
quantization does not cause a huge auditory problem, so it will be
best to just perform the scale and shift instead of just the shift.